refactor this to a concern so I can reuse it when inheriting from another controller
