+++ app/controllers/rad_common/attachments_controller.rb
@@ -2,12 +2,10 @@ module RadCommon
   class AttachmentsController < ApplicationController
     require 'open-uri'
 
-    include ActiveStorage::Downloading
+    include RadCommon::ActiveStorageDownloader
 
     before_action :authenticate_user!, only: :destroy
     before_action :set_variant, only: :download_variant
-
-    attr_reader :blob
 
     def download
       skip_authorization
@@ -59,34 +57,6 @@ module RadCommon
         rescue NoMethodError
           @variant = nil
         end
-      end
-
-      def serve_active_storage_file(attachment, filename)
-        # TODO: refactor this to use single method for production and development, may need to wait until Rails 6
-        if Rails.env.production?
-          serve_file attachment.service_url,
-                     attachment.blob.filename.extension_with_delimiter,
-                     attachment.blob.content_type,
-                     filename
-        else
-          @blob = attachment.blob
-          download_and_serve_file attachment.blob.filename.extension_with_delimiter,
-                                  attachment.blob.content_type,
-                                  filename
-        end
-      end
-
-      def download_and_serve_file(ext, content_type, filename)
-        # does not support variants/resizing
-        download_blob_to_tempfile do |file|
-          send_data file.read, filename: "#{filename}.#{ext}", type: content_type, disposition: 'inline'
-        end
-      end
-
-      def serve_file(url, ext, content_type, filename)
-        # crashes locally, only use for production
-        data = URI.open(url)
-        send_data data.read, filename: "#{filename}.#{ext}", type: content_type, disposition: 'inline'
       end
   end
 end
